# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: API Integration
# during dev, should point to your local machine
# host: localhost:5040
host: 52.39.212.226:5054
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/getRoles:
    # binds a127 app logic to a route
    x-swagger-router-controller: roles_ctrl
    get:
      tags: ['Role Module']
      description: Get All roles
      summary: Get All roles
      # used as the method name of the controller
      operationId: getRoles  
      parameters:    
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
...
  /api/roles/update_permission/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: roles_ctrl
    put:
      tags: ['roles']
      description: update role permission
      summary : Update role permission
      # used as the method name of the controller
      operationId: updatePermission
      parameters:
        - name: id
          in: path
          description: role id
          required: true
          type: string        
        - name: body
          in: body
          description: Body
          required: false
          schema:
            # a pointer to a definition
            $ref: '#/definitions/updatePermission'   
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string           
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"         
...
  /api/roles:
    # binds a127 app logic to a route
    x-swagger-router-controller: roles_ctrl
    get:
      tags: ['roles']   
      description: Get roles
      summary: Get roles
      # used as the method name of the controller
      operationId: getRoles
      parameters:
        - name: query
          in: query
          description: query
          required: false
          type: string
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"     
    post:
      tags: ['roles']
      description: Add role
      summary: Add role
      # used as the method name of the controller
      operationId: addRole
      parameters:
        - name: body
          in: body
          description: user infomation
          schema:
           $ref: '#/definitions/addRole'
          required: true
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/roles/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: roles_ctrl
    get:
      tags: ['roles']
      description: Get role by id
      summary: Get role 
      # used as the method name of the controller
      operationId: getRoleById
      parameters:
        - name: id
          in: path
          description: role id
          required: true
          type: string
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags: ['roles']
      description: update role
      summary : Update role
      # used as the method name of the controller
      operationId: updateRole
      parameters:
        - name: id
          in: path
          description: role id
          required: true
          type: string        
        - name: body
          in: body
          description: Body
          required: false
          schema:
            # a pointer to a definition
            $ref: '#/definitions/addRole'   
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string     
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"     
    delete:
      tags: ['roles']
      description: delete role
      summary : Delete role
      # used as the method name of the controller
      operationId: deleteRole
      parameters:
        - name: id
          in: path
          description: role id
          required: true
          type: string
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:                         
...
  addRole:
    type: object
    properties:
      title:
        type: string
      description:
        type: string
  updatePermission:
    type: object
    properties:
      permission_id:
       type: array
       items:
        type: string
...
  appResponse:
    # required:
    #   - code
    properties:
      code:
        type: number
      status:
        type: string        
      message:
        type: string        
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  # by sunny on 15-april-2017
...